ml/requirements.txt
pytesseract==0.3.10
Pillow==10.0.0
numpy==1.24.3
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
opencv-python==4.8.1.78
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
